switch cooldowns back to enums

+++ StreamSchedule/Commands/EvaluateUsers.cs
@@ -8,7 +8,7 @@ internal class EvaluateUsers : Command
     internal override string Call => "evaluate";
     internal override Privileges MinPrivilege => Privileges.Uuh;
     internal override string Help => "evaluate user's scores and assign privileges accordingly: [username](optional) ";
-    internal override TimeSpan Cooldown => TimeSpan.FromSeconds(Cooldowns.Long);
+    internal override TimeSpan Cooldown => TimeSpan.FromSeconds((int)Cooldowns.Long);
     internal override Dictionary<string, DateTime> LastUsedOnChannel { get; set; } = [];
     internal override string[]? Arguments => ["s"];
