+++ StreamSchedule/Commands/Helps.cs
@@ -7,7 +7,7 @@ internal class Helps : Command
     internal override string Call => "helps";
     internal override Privileges MinPrivilege => Privileges.None;
     internal override string Help => "show command help: [command name] ";
-    internal override TimeSpan Cooldown => TimeSpan.FromSeconds(Cooldowns.Medium);
+    internal override TimeSpan Cooldown => TimeSpan.FromSeconds((int)Cooldowns.Medium);
     internal override Dictionary<string, DateTime> LastUsedOnChannel { get; set; } = [];
     internal override string[]? Arguments => null;
