+++ StreamSchedule/Commands/Ping.cs
@@ -7,7 +7,7 @@ internal class Ping : Command
     internal override string Call => "ping";
     internal override Privileges MinPrivilege => Privileges.Trusted;
     internal override string Help => "ping";
-    internal override TimeSpan Cooldown => TimeSpan.FromSeconds(Cooldowns.Minute);
+    internal override TimeSpan Cooldown => TimeSpan.FromSeconds((int)Cooldowns.Minute);
     internal override Dictionary<string, DateTime> LastUsedOnChannel { get; set; } = [];
     internal override string[]? Arguments => null;
